// Dashboard-Frontend/src/app/models/salaryTime.model.ts
export class salaryTimeModel {

  id?: number;
  areaSalaryId?: number;
  salary?: number;
  expiredMonth?: string;
  staffType?: string;
  createdDate?: string;
  createdUser?: string;
  updatedDate?: string;
  updatedUser?: string;
  expiredMilis?: number;
  areaCode?: string;

  constructor() {}
}